requirements.txt
streamlit==1.31.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2
supabase==2.3.4
numpy==1.26.4
